Refacto(history): Fix linter warning
