+++ Makefile
@@ -26,7 +26,8 @@ SRC	=	src/main.c										\
 		src/ncurse/my_put.c								\
 		src/ncurse/set_get_term_name.c					\
 		src/ncurse/start_ncurses.c						\
-		src/ncurse/line_edition/line_edition.c			\
+		src/ncurse/line_edition/move_in_history.c		\
+		src/ncurse/get_string.c							\
 		src/parser/parser.c								\
 		src/parser/get_token.c							\
 		src/parser/token_dup.c							\
