+++ src/ncurse/line_edition/move_in_history.c
@@ -11,7 +11,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-static char *move_in_history(char *save, int *index, char **history, int ch)
+char *move_in_history(int ch, char *save, int *index, char **history)
 {
     int len = my_arraylen(history) - 1;
     char *line = NULL;
@@ -23,7 +23,7 @@ static char *move_in_history(char *save, int *index, char **history, int ch)
             (*index)--;
             line = strdup(history[*index]);
         }
-    } if (ch == KEY_DOWN) {
+    } else {
         if (*index >= len && save != NULL)
             line = save;
         if (*index < len) {
@@ -32,19 +32,4 @@ static char *move_in_history(char *save, int *index, char **history, int ch)
         }
     }
     return line;
-}
-
-char *line_edition(int ch, char *save, int *index, char **history)
-{
-    char *line = NULL;
-
-    if (ch == KEY_UP || ch == KEY_DOWN)
-        line = move_in_history(save, index, history, ch);
-    return line;
-}
-
-void free_line_edition(char **history, char *save)
-{
-    destroy_array(history);
-    free(save);
 }
